+++ pyqtgraph_scope_plots/__init__.py
@@ -24,11 +24,10 @@ from .enum_waveform_plotitem import EnumWaveformPlot
 from .plots_table_widget import PlotsTableWidget
 from .signals_table import (
     DeleteableSignalsTable,
-    HasDataSignalsTable,
-    StatsSignalsTable,
     ColorPickerSignalsTable,
     DraggableSignalsTable,
 )
+from .stats_signals_table import StatsSignalsTable
 from .timeshift_signals_table import TimeshiftSignalsTable
 from .transforms_signal_table import TransformsSignalsTable
 from .search_signals_table import SearchSignalsTable
