+++ pyqtgraph_scope_plots/csv/csv_plots.py
@@ -45,7 +45,8 @@ from ..multi_plot_widget import MultiPlotWidget
 from ..plots_table_widget import PlotsTableWidget
 from ..save_restore_model import BaseTopModel, HasSaveLoadDataConfig
 from ..search_signals_table import SearchSignalsTable
-from ..signals_table import ColorPickerSignalsTable, StatsSignalsTable
+from ..signals_table import ColorPickerSignalsTable
+from ..stats_signals_table import StatsSignalsTable
 from ..time_axis import TimeAxisItem
 from ..timeshift_signals_table import TimeshiftSignalsTable
 from ..transforms_signal_table import TransformsSignalsTable
