+++ pyqtgraph_scope_plots/xy_plot.py
@@ -94,13 +94,6 @@ class XyPlotWidget(BaseXyPlot, pg.PlotWidget):  # type: ignore[misc]
             self._update()
         self.sigXyDataItemsChanged.emit()
 
-    def _get_region(self) -> Tuple[float, float]:
-        """Gets the currently selected region from the plot, or (-inf, inf) by default."""
-        if isinstance(self._plots, LinkedMultiPlotWidget) and isinstance(self._plots._last_region, tuple):
-            return self._plots._last_region
-        else:
-            return (-float("inf"), float("inf"))
-
     @staticmethod
     def _get_correlated_indices(
         x_ts: npt.NDArray[np.float64], y_ts: npt.NDArray[np.float64], start: float, end: float
@@ -124,7 +117,7 @@ class XyPlotWidget(BaseXyPlot, pg.PlotWidget):  # type: ignore[misc]
         for data_item in self.listDataItems():  # clear existing
             self.removeItem(data_item)
 
-        region = self._get_region()
+        region = HasRegionSignalsTable._region_of_plot(self._plots)
         data = self._plots._data
         for x_name, y_name in self._xys:
             x_ts, x_ys = data.get(x_name, (None, None))
